Fix: add manage message function
